requirement.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blessed==1.19.1
bpython==0.23
cchardet==2.1.7
celery==5.2.7
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
clevercsv==0.7.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
curtsies==0.4.1
cwcwidth==0.1.8
Deprecated==1.2.13
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-timezone-field==5.0
djangorestframework==3.14.0
greenlet==2.0.1
idna==3.4
kombu==5.2.4
numpy==1.23.5
packaging==21.3
pandas==1.5.1
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pycodestyle==2.10.0
Pygments==2.13.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
pyxdg==0.28
redis==4.3.4
regex==2022.10.31
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tabview==1.4.4
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wilderness==0.1.8
wrapt==1.14.1
